+++ aioutils/__init__.py
@@ -1,5 +1,6 @@
 from .pool import Pool, Group
 from .bag import Bag, OrderedBag
+from .yielder import Yielder, OrderedYielder
 
-__all__ = ['Pool', 'Group', 'Bag', 'OrderedBag']
-__version__ = '0.2.1'
+__all__ = ['Pool', 'Group', 'Bag', 'OrderedBag', 'Yielder', 'OrderedYielder']
+__version__ = '0.3.0'